chore(toolbox-core): Fix deadlock issue in e2e tests
