+++ packages/toolbox-core/tests/test_sync_client.py
@@ -14,7 +14,6 @@
 
 
 import inspect
-from asyncio import run_coroutine_threadsafe
 from typing import Any, Callable, Mapping, Optional
 from unittest.mock import AsyncMock, Mock, patch
 
@@ -27,31 +26,6 @@ from toolbox_core.sync_client import ToolboxSyncClient
 from toolbox_core.sync_tool import ToolboxSyncTool
 
 TEST_BASE_URL = "http://toolbox.example.com"
-
-
-# The original `ToolboxSyncClient.close()` is a blocking method because it
-# calls `.result()`. In the pytest environment, this blocking call creates a
-# deadlock during the test teardown phase when it conflicts with the
-# `sync_client_environment` fixture that also manages the background thread.
-#
-# By replacing `close` with our non-blocking version for the test run,
-# we prevent this deadlock and allow the test suite to tear down cleanly.
-@pytest.fixture(autouse=True)
-def patch_sync_client_for_deadlock(monkeypatch):
-    """
-    Automatically replaces the blocking `ToolboxSyncClient.close()` method
-    with a non-blocking version for the entire test run.
-    """
-
-    def non_blocking_close(self):
-        """A replacement for close() that doesn't block."""
-        loop = self.__class__._ToolboxSyncClient__loop
-        async_client = self._ToolboxSyncClient__async_client
-
-        coro = async_client.close()
-        run_coroutine_threadsafe(coro, loop)
-
-    monkeypatch.setattr(ToolboxSyncClient, "close", non_blocking_close)
 
 
 @pytest.fixture
